feat(html): Add header block with logo and links
This introduces a new header block to the project that consist of two
main elements: a logo and a navigation menu.

- Logo: Inside the header, a "header__logo" div has been added,
displaying the text "Logo."

- Navigation Menu: The header also includes a navigation menu, created
using an unordered list (ul) with list items (li) representing each
navigation link. Each link is denoted by the "header__nav-item" class,
and within each link, an anchor (a) element is used with the
"header__nav-link" class.
_